:hammer: [Protocol Types] Improved any/3 and list/1 type
Lists were accepting both `enum/1` and `any/2` types, while not
supporting them. With this commit `enum/1` is now supported, while
`any/2` it is explicitly said it won't support it (at least for now).

Also I moved the `field` option for `any/2` into the actual type. This
states the mandatory nature of the the, now, `any/3` type.